+++ src/test/java/org/seasar/doma/wrapper/FloatWrapperTest.java
@@ -8,13 +8,13 @@ public class FloatWrapperTest extends TestCase {
   public void testIncrement() {
     var wrapper = new FloatWrapper(10f);
     wrapper.increment();
-    assertEquals(Float.valueOf(11f), wrapper.get());
+    assertEquals(11f, wrapper.get());
   }
 
   /** */
   public void testDecrement() {
     var wrapper = new FloatWrapper(10f);
     wrapper.decrement();
-    assertEquals(Float.valueOf(9f), wrapper.get());
+    assertEquals(9f, wrapper.get());
   }
 }
